+++ src/CreditCardGateway.php
@@ -8,7 +8,7 @@ use Pronamic\WordPress\Pay\Plugin;
 /**
  * Title: WooCommerce Credit Card gateway
  * Description:
- * Copyright: 2005-2019 Pronamic
+ * Copyright: 2005-2020 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma